Fix bugs found in comprehensive code audit
- Restore missing ban check in handshake (banned players could
  reconnect freely)
- Reset session to waiting-for-drivers when last client disconnects
- Fix standings_seq type to uint32_t matching wire format (was
  uint16_t, would wrap after 65535 changes)
- Add session_index bounds check in LAN discovery response
- Remove dead duplicate return in build_welcome_trailer
- Fix probe stack buffer overflow with long args (2048 -> 4096)

+++ accd/tick.c
@@ -256,7 +256,7 @@ done:
 void
 tick_run(struct Server *s)
 {
-	uint16_t *last_standings_seq = &s->session.last_standings_seq;
+	uint32_t *last_standings_seq = &s->session.last_standings_seq;
 	uint8_t *last_phase = &s->session.last_phase;
 
 	s->tick_count++;
